feat(web-core): enhance form select
