+++ @xen-orchestra/web-core/lib/packages/form-select/index.ts
@@ -1,5 +1,2 @@
-export * from './type.ts'
-export * from './use-form-option.ts'
-export * from './use-form-options.ts'
+export * from './types.ts'
 export * from './use-form-select.ts'
-export * from './use-form-select-keyboard-navigation.ts'
